+++ CHANGELOG.md
@@ -4,11 +4,21 @@ All notable changes to the "kaas-vscode" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.10] - August 6, 2024
 
 ### New Features
 
 - Support to display job/report/cache pages directly in the vscode extension.
+- Optimize validation checks to run once per workspace before test execution instead of per test.
+- Improve the user experience.
+
+### Bug Fixes
+
+- Fix test execution in multi-workspace setups to use correct workspace folder and git information.
+- Fix `--match-test` flag to use correct regular expression for function names.
+- Fix test run timer not stopping when jobs complete successfully or fail.
+- Fix error handling when clicking on non-executable test container nodes.
+- Fix displaying the contract and test names in the test explorer.
 
 ## [0.0.9] - July 24 2025